refactor: remove comparsion
