+++ src/router.ts
@@ -23,7 +23,6 @@ import Dashboard from './views/containers/dashboard.vue';
 import Trace from './views/containers/trace.vue';
 import Topology from './views/containers/topology/topology.vue';
 import Alarm from './views/containers/alarm.vue';
-import Comparison from './views/containers/comparison.vue';
 import Profile from './views/containers/profile.vue';
 
 Vue.use(Router);
@@ -58,10 +57,6 @@ const router = new Router({
         {
           path: 'alarm',
           component: Alarm,
-        },
-        {
-          path: 'comparison',
-          component: Comparison,
         },
         {
           path: 'profile',
